refactor: de-duplicate the overflow-drain helper across the two attach tests
The timeout-guarded pull-until-overflow loop was copy-pasted in the in-process
and real-socket overflow tests, differing only in the laggard-frame bound. Hoist
it to contractCorpus.testlib.ts (drainForOverflow), shared by both. (/simplify)
